% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_elements.R
\docType{class}
\name{Capacitor}
\alias{Capacitor}
\title{Capacitor Class}
\description{
Class for the capacitors.
}
\details{
Capacitor class
}
\section{Methods}{

\describe{
\item{num_impedance compute the impedance considering the value of the capacitance}
\item{sym_impedance give the symbolic impedance considering the name of the capacitor}
}
}

\section{Super class}{
\code{\link[ltspicer:ElectricalObject]{ltspicer::ElectricalObject}} -> \code{Capacitor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Capacitor$new()}}
\item \href{#method-num_impedance}{\code{Capacitor$num_impedance()}}
\item \href{#method-sym_impedance}{\code{Capacitor$sym_impedance()}}
\item \href{#method-set_capacitance}{\code{Capacitor$set_capacitance()}}
\item \href{#method-spice_line}{\code{Capacitor$spice_line()}}
\item \href{#method-clone}{\code{Capacitor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Capacitor$new(capacitor_txt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_impedance"></a>}}
\if{latex}{\out{\hypertarget{method-num_impedance}{}}}
\subsection{Method \code{num_impedance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Capacitor$num_impedance(f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sym_impedance"></a>}}
\if{latex}{\out{\hypertarget{method-sym_impedance}{}}}
\subsection{Method \code{sym_impedance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Capacitor$sym_impedance(f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_capacitance"></a>}}
\if{latex}{\out{\hypertarget{method-set_capacitance}{}}}
\subsection{Method \code{set_capacitance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Capacitor$set_capacitance(new_cap)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spice_line"></a>}}
\if{latex}{\out{\hypertarget{method-spice_line}{}}}
\subsection{Method \code{spice_line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Capacitor$spice_line()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Capacitor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
