% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_elements.R
\docType{class}
\name{ElectricalObject}
\alias{ElectricalObject}
\title{ElectricalObject}
\description{
A base class for the electrical objects used in LTspice netlist.
}
\details{
ElectricalObject class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the object}

\item{\code{plus}}{name of the plus node}

\item{\code{minus}}{name of the minus node}

\item{\code{value}}{value of the electrical object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{value}}{value of the electrical object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-spice_line}{\code{ElectricalObject$spice_line()}}
\item \href{#method-clone}{\code{ElectricalObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spice_line"></a>}}
\if{latex}{\out{\hypertarget{method-spice_line}{}}}
\subsection{Method \code{spice_line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectricalObject$spice_line()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectricalObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
