% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_utils.R
\name{parse_netlist}
\alias{parse_netlist}
\title{Parse LTspice netlist}
\usage{
parse_netlist(netlist_path)
}
\arguments{
\item{netlist_path}{string, path to ltspice netlist}
}
\value{
string vector with parse lines of the netlist
}
\description{
A one-liner to parse the netlist given the path to it.
}
\examples{
netlist <- parse_netlist('spice_files/2m_cable_model.net')
print(netlist)

}
