% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_utils.R
\name{parse_number_suffix}
\alias{parse_number_suffix}
\title{Parse number with appropriate suffix}
\usage{
parse_number_suffix(num_with_suffix)
}
\arguments{
\item{num_with_suffix}{string}
}
\value{
double, appropriate number
}
\description{
Parse number with appropriate suffix
}
\examples{
parse_number_suffix('45     u')
parse_number_suffix('45Megkfdsaf')
parse_number_suffix('4.623M') == parse_number_suffix('4.623m')
}
