% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice_utils.R
\name{part_impedance}
\alias{part_impedance}
\title{Impedance}
\usage{
part_impedance(electrical_object)
}
\arguments{
\item{electrical_object}{the object inheriting from the class ElectricalObject}
}
\value{
string, ltspice command measuring the impedance
}
\description{
Impedance
}
