% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_utils.R
\name{read_raw_spice}
\alias{read_raw_spice}
\title{Read raw LTspice file}
\usage{
read_raw_spice(path, ascii = TRUE)
}
\arguments{
\item{path}{path to the raw file}

\item{ascii}{boolean, are the values wrote in ascii?}
}
\value{
a tibble of data traces in raw file
}
\description{
Read raw LTspice file
}
