% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_elements.R
\docType{class}
\name{Resistor}
\alias{Resistor}
\title{Resistor Class}
\description{
Class for the resistors
}
\details{
Resistor class
}
\section{Methods}{

\describe{
\item{num_impedance compute the impedance considering the value of the resistance}
\item{sym_impedance give the symbolic impedance considering the name of the resistor}
\item{set_resistance set the resistance}
\item{spice_line produce the line for the netlist}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Resistor$new()}}
\item \href{#method-set_value}{\code{Resistor$set_value()}}
\item \href{#method-spice_line}{\code{Resistor$spice_line()}}
\item \href{#method-num_impedance}{\code{Resistor$num_impedance()}}
\item \href{#method-sym_impedance}{\code{Resistor$sym_impedance()}}
\item \href{#method-clone}{\code{Resistor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resistor$new(resistor_txt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resistor$set_value(new_res)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spice_line"></a>}}
\if{latex}{\out{\hypertarget{method-spice_line}{}}}
\subsection{Method \code{spice_line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resistor$spice_line()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_impedance"></a>}}
\if{latex}{\out{\hypertarget{method-num_impedance}{}}}
\subsection{Method \code{num_impedance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resistor$num_impedance(f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sym_impedance"></a>}}
\if{latex}{\out{\hypertarget{method-sym_impedance}{}}}
\subsection{Method \code{sym_impedance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resistor$sym_impedance(f)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resistor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
