% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_utils.R
\name{write_netlist}
\alias{write_netlist}
\title{Write netlist to a file}
\usage{
write_netlist(netlist_lines, netlist_path)
}
\arguments{
\item{netlist_path}{}
}
\value{

}
\description{
Write netlist to a file
}
